adds helper method to normalize keys
