+++ tests/Symfony/Tests/Component/DependencyInjection/Extension/ExtensionTest.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Tests\Component\DependencyInjection\Extension;
 
+use Symfony\Component\DependencyInjection\Extension\Extension;
+
 require_once __DIR__.'/../Fixtures/includes/ProjectExtension.php';
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -34,5 +36,39 @@ class ExtensionTest extends \PHPUnit_Framework_TestCase
 
         $extension->load('bar', array('foo' => 'bar'), $config = new ContainerBuilder());
         $this->assertEquals(array('project.parameter.bar' => 'bar', 'project.parameter.foo' => 'bar'), $config->getParameterBag()->all(), '->load() calls the method tied to the given tag');
+    }
+
+    /**
+     * @dataProvider getKeyNormalizationTests
+     */
+    public function testNormalizeKeys($denormalized, $normalized)
+    {
+        $this->assertSame($normalized, Extension::normalizeKeys($denormalized));
+    }
+
+    public function getKeyNormalizationTests()
+    {
+        return array(
+            array(
+                array('foo-bar' => 'foo'),
+                array('foo_bar' => 'foo'),
+            ),
+            array(
+                array('foo-bar_moo' => 'foo'),
+                array('foo-bar_moo' => 'foo'),
+            ),
+            array(
+                array('foo-bar' => null, 'foo_bar' => 'foo'),
+                array('foo-bar' => null, 'foo_bar' => 'foo'),
+            ),
+            array(
+                array('foo-bar' => array('foo-bar' => 'foo')),
+                array('foo_bar' => array('foo_bar' => 'foo')),
+            ),
+            array(
+                array('foo_bar' => array('foo-bar' => 'foo')),
+                array('foo_bar' => array('foo_bar' => 'foo')),
+            )
+        );
     }
 }
